Better handling of inlined files
This restores the previous handling of inlined files, which was a strict
requirement in earlier versions, now more of an optimization.

littlefs now tries to keep small inlinable files in RAM. This doesn't
always work because inlined files can be quite large, so littlefs only
tries if it knows the inlineable file fits in the file's buffer.

I had a bit of a hard time working truncate/fruncate into the scheme. It
seems simple on paper: if truncate/frunate make a file "small", move it
into our buffer. But if we already have data in our buffer, we can't make
the file small without potentially clobbering the buffer. This conflicts
with error recovery and leaves us in a bit of a bind.

Two options:

1. Flush the file on truncate/fruncate so we can use the buffer for
   caching small files.

   This risks unnecessary disk writes. Consider fruncating a small
   log you know fits in the file buffer.

2. Allow small files to sometimes not be cached in the buffer. Fix this
   state on sync.

   This risks bugs caused by relying on the sometimes-incorrect small
   file invariant.

   This can still also cause unnecessary disk writes if you, say, write
   to the file between truncate and sync.

This goes with the option 2. I think option 1 has more failure cases
that can be problematic and option 2 allows us to avoid disk writes in
more cases. But it isn't clear what the best option is.